Change the Data Type of 'flags' in the Environment
When building for a 64 bit target, unsigned long is
64 bits instead of 32 bits.  The checksum is still
calculated assuming it is 32 bits.  This commit
changes flags to 'unsigned int'.

+++ include/environment.h
@@ -156,7 +156,7 @@ extern char *env_name_spec;
 typedef struct environment_s {
 	uint32_t	crc;		/* CRC32 over data bytes	*/
 #ifdef CONFIG_SYS_REDUNDAND_ENVIRONMENT
-	unsigned long	flags;		/* active/obsolete flags	*/
+	unsigned int	flags;		/* active/obsolete flags	*/
 #endif
 	unsigned char	data[ENV_SIZE]; /* Environment data		*/
 } env_t
